+++ libs/vision/include/mrpt/vision/types.h
@@ -329,16 +329,13 @@ struct TImageROI
 {
 	// Constructors
 	TImageROI() = default;
-	TImageROI(float x1, float x2, float y1, float y2);
+	TImageROI(size_t x1, size_t x2, size_t y1, size_t y2);
 
-	// Members
-	/** X coordinate limits [0,imageWidth)
-	 */
-	float xMin{0}, xMax{0};
-	/** Y coordinate limits [0,imageHeight)
-	 */
-	float yMin{0}, yMax{0};
-};  // end struct TImageROI
+	/** X coordinate limits [0,imageWidth) */
+	size_t xMin{0}, xMax{0};
+	/** Y coordinate limits [0,imageHeight) */
+	size_t yMin{0}, yMax{0};
+};
 
 /** A structure containing options for the matching
  */
